+++ spec/app/models/word.rb
@@ -1,7 +1,10 @@
 class Word
   include Mongoid::Document
   field :name, :type => String
+
   belongs_to :dictionary
+
   embeds_many :definitions, :validate => false
+  embeds_many :word_origins, :validate => false
   embeds_one :pronunciation, :validate => false
 end